fix: handle Buffer serialization
Buffer has a toJSON implementation which is called before our replacer function when serialising. It's JSON form is an object like { type: 'Buffer, data: number[] }`.

This gets mangled in storage and does not get deserilaised in the reviver. I think we'd rather have UintA8Arrays everywhere so this PR fixes it so that it gets revived as that.

License: MIT
Signed-off-by: Oli Evans <oli@protocol.ai>