feat: sync encode/decode delegations (#276)
- add methods toBytes and fromBytes useful to store as bytes instead of
strings
- `encodeDelegations` is now `delegationsToString``
- `decodeDelegations` is now `stringToDelegations``
- all functions are sync now

Writing delegations to a car uses `@ipld/car/buffer-writer` now to make
it sync, reading from a CAR it's more complicated 🤫.
`ipld/car` does not support sync reading so this PR includes a refactor
of `CarReader` to be sync, this is a temporary solution until i can get
a proper PR to `ipld/car` to add support for sync reading. It's not
trivial to do so because of the way ipld/car code is structured and i
dont want to just add a sync copy/paste of the decoder functions to it.