feat: add `redirect` field to voucher `nb`s
We'd like to allow clients (especially on the web) to specify a redirect when registering a new space. To do this we add a `redirect` field to `voucher/claim` and `voucher/redeem` and
use it to decide whether to send a user to the normal landing page or to redirect them to the specified URL after claiming and redeeming a voucher.