feat: add `prev` caveat to `store/list` and `upload/list`
Per https://github.com/web3-storage/w3ui/issues/143 we'd like to give
clients a way to get the previous page of results rather than the next
page.

This change introduces a `prev` caveat which, when set to true, will
return the page of results preceding `cursor`.

I considered calling this `reverse` or `rev` but I think that implies
that results will be returned sorted in reverse, which is not
the intent.