chore: small readme updates (#421)
This fixes the capability imports in the upload-client examples to close
#326

I didn't see any more references to the old import paths, but lmk if I
missed any.

I also changed the casing of `Agent` and `Space` in a few places; when I
first wrote the access client readme I went a little nuts with the
capitalization, but later decided it looked weird and stopped
capitalizing them everywhere. So this just makes things consistent with
the docs site, etc.