feat: add access-api + access-client agent test and improve types (#533)
Motivation:
* there weren't any tests that test access-api + access-client agent, so
this adds a stub for one
* it includes a lot of type adjustments that should make it easier to
write more tests in followup
* specifically it removed some old types I had tried to make when
testing access delegate. I added a new `ServiceInvoke` type that is
better than the ones I had made before. This is useful for making an
'invoke' function that will handle a subset of the caps available on
`Service` but expect the same success/fail types as Service