feat: move validation flow to a Durable Object to make it ⏩ fast ⏩ fast ⏩ fast ⏩  (#449)
Introduce a [Cloudflare Durable
Objects](https://developers.cloudflare.com/workers/runtime-apis/durable-objects/)-based
space verification workflow.

Rather than stashing a UCAN in KV and polling inside the Websocket
handler until eventual consistency makes it available to the
registration process in the CLI or w3console, we use a Durable Object to
create a direct connection between the Websocket and the HTTP POST that
provides the delegation.

It feels most natural to create a Durable Object per space and forward
the Websocket request on to the Durable Object and let it handle it
entirely. This necessitates adding the space DID information to the HTTP
path, which means adding a new endpoint and will require clients to be
updated in the wild before we get rid of the KV flow entirely. We should
remove KV-based verification once this has rolled out to all of our
users.

TODO
- [x] write tests
- [ ] make sure error handling makes sense

---------

Co-authored-by: Benjamin Goering <171782+gobengo@users.noreply.github.com>