fix: Agent#authorize races websocket and polling access/claim  (#560)
TIL:
* the websocket definitely doesn't really work in access-api tests,
because access-ws is an entirely different cf worker! it's not running
in those access-api tests
* so 100% of the time the websocket immediately terminates with
ECONNREFUSED, but afaict that error isn't really bubbled up when Agent
calls `await ws.open`. This still doesn't bubble up that error, but does
allow for that open to be abortable
* and then authorize() races that websocket and polling access claim. If
access/claim polling wins, it aborts the websocket, and everything kinda
maybe works

End result
* this re-enables a test against Agent#authorize, which passes due to
the polling on access/claim
  * in the wild maybe the websocket will win the race sometimes

(so it can be used in tests)