fix: set currentSpace inside addSpace if currentSpace isn't set
I noticed this wasn't happening because w3ui's w3console was in a state where the space picker was populated with many spaces, but currentSpace wasn't set.

I think we should avoid this situation at this level because it's most convenient for me, but open to moving this assurance to another level if.