chore: set package versions manually for now (#586)
Per instructions and examples here:


https://github.com/googleapis/release-please/blob/main/docs/manifest-releaser.md

Set package numbers manually for now. Release Please is currently very
confused because we set a prerelease version number using a commit and
now it wants to release everything with that version number. This should
fix that.

After we release these packages we should be able to delete these and it
will go back to guessing based on conventional commits.