fix: call claimAccess at the end of authorizeWithPollClaim
I'm testing out using `authorizeWithPollClaim` instead of `authorizeWithSocket` to see if it will be more responsive and therefore better for demos. It worked, but doesn't currently run `claimDelegations` and therefore doesn't infer spaces from the list of claimed delegations.

This feels a bit duplicative because we're already calling it several times, but one more time won't hurt much and this makes it more consistent with `authorizeWithSocket`.