feat: get `access/claim` authorization wait function working (#666)
Build on @bengo's `access/claim`-polling-based
`waitForDelegationOnSocket` replacement. This will let us replace the
websocket-based authorization wait function which makes the UX much more
responsive - rather than waiting an indeterminate amount of time after
clicking the email link, the process completes almost immediately with
this change.

The wait function is also pluggable so we that clients can implement
different polling strategies, which should make it easier to test and
use a receipt-based version of this in the future.

Added tests for this that use `sinon` to mock out service handlers and
ensure they are called the right number of times.