feat: allow clients to pass uploadFile an onUploadProgress callback
use the ipfs-utils fetch implementation to enable upload progress reporting

I'm trying to get this to work with w3console, but so far always see `loaded` and `total` come back the same, so I'm not 100% sure this is working as expected.

I suspect the issue may be related to an error about trying to set the Content-Length header on the request. This is doomed because either:

a) CORS configuration disallows setting Content-Length on cross-domain requests or
b) (I suspect) a Content-Length header was not included in the request that was signed by the S3 upload tools, so AWS just hangs when we try to actually do the PUT

Open questions:
How hard would it be to add a Content-Length header to whatever we are using to generate the signed URL?