fix: account for edge cases in polling abort
My previous implementation relied on canceling a timeout to stop the polling process. This had a number of edge cases, including one where the signal was aborted before first round of polling happened, which is quite common since we do some network calls before we even get to this code.

Stop worrying about canceling the timeout since we'll just stop polling whenever we execute `poll` in this new implementation - generally this relies on a piece of persistent state rather than the assumption that the correct timeout will get canceled, so is quite a bit more robust.