er-add and tests after w3infra work
we had some tests that said a space couldn't have two providers and some that said they could. my understanding is that we do not want to support multiple providers on a space, so update tests accordingly.

these changes flowed out of the dynamodb implementation and things I learned while writing that and generally make our mock provisions storage more consistent with that implementation