feat: w3 aggregate protocol client and api implementation (#787)
This PR is a first iteration for w3 aggregate client and API
implementations. It follows same patterns we identified while working on
`upload-api` and `upload-client` and is agnostic from the deployment
service. `aggregate-api` exposes a lib and a set of tests that
implementers with infra should follow, as well as the needed interfaces
that must be fulfilled. It also includes `fx` usage for
`aggregate/offer` return value

It follows spec
https://github.com/web3-storage/specs/blob/feat/filecoin-spec/w3-aggregation.md

There are two interfaces defined in this service:
- `offerStore` - responsible for receiving offers to be queued.
Implementer receives `commD` of aggregate together with Offer content.
These will be used to give SPs. Note that implementer should keep track
of the offer's `commD` so that it can query over time when a deal is
done (or failures), so that the `fx` invocation can be executed and a
receipt generated
- `aggregateStore` - store of aggregates already places with SPs
(implementor like `spade-proxy` will have this store hooked with spade
DB)

New detailed issues for above will be created before this PR gets
merged.

Part of  https://github.com/web3-storage/w3filecoin/issues/19

Closes #772 and #773