fix: per PR feedback use allocate in upload/add for consistency
this ensures people can't upload to a blocked space, and makes it more consistent.

this also gets rid of the `size` parameter in `space/allocate` because:

a) we weren't using it for anything
b) we don't have size information in `upload/add`

I think we'll want to add this back in later, possibly as an optional parameter that triggers a check against provisioned allocation, or we'll want to split space/allocate into two capabilities.