fix: manually typecast getSpaceInfo result
I fought with TS all day and chatted with @gozala near the end of it. We discovered that the types `derive` returns don't seem to play nice with the type inference we're doing on the result of `invokeAndExecute`. He has some longer-term fixes to Ucanto that should help, so I'm opting to just do a little gentle typecasting to fix this for now.