fix: type error in access client types
This has been red for a bit, but I assumed it was superfluous until I saw the error in a downstream project I was working on earlier today.

I don't fully understand why this fixes the error I was seeing, but a) it does and b) I don't see much value in the generic type here.

Pushing this up to start a discussion about whether this is an acceptable way to fix this issue - would love any thoughts on a better way to fix if anyone has ideas!