feat!: add size to `store/remove` receipt (#969)
This PR adds a `size` property to the `store/remove` receipt. This is so
we can account for removals properly.

This also moves the `store/*` and `upload/*` receipt types into the
`@web3-storage/capabilities` package to be consistent with the rest of
them. This accounts for much of the churn here.

TODO: update
https://github.com/web3-storage/w3infra/blob/9f0c50dd2cf4adec616483d7c8d29cfdec03129c/ucan-invocation/functions/space-metrics-store-remove-size-total.js#L43-L51
(in separate PR)

BREAKING CHANGE: Returning the `size` means that we need to fetch the
stored item beforehand, and if it does not exist throw a
`StoreItemNotFound` error. This is a change from the current behaviour
which returns successfully even if the item is not present in the space.