feat: add `store/get` and `upload/get` capabilities (#942)
Capability definitions to allow clients to check for inclusion of cid in
store shards or upload roots. Also allows client to fetch the additional
properties we store. This will, for example, let a client verify all
shards for an upload have been removed before removing an upload.

_I'm aiming to simplify things with this PR so that we return insertedAt
info on the Upload/Store ListItem types, and return the same object
shape for a single object get. If this is pleasing, then I will continue
in this direction and make Remove also return the same shape if it
removes something_

TODO:
- [x] spec update https://github.com/web3-storage/specs/pull/82
- [x] tests


License: MIT

---------

Signed-off-by: Oli Evans <oli@protocol.ai>
Co-authored-by: Alan Shaw <alan.shaw@protocol.ai>