feat: optionally require plans for provisioning
We need this to ship billing, to stop users without a payment provider plan from provisioning new spaces.

I've added a flag to the `ProviderServiceContext` that lets us disable this requirement - this is extremely useful in testing and will be helpful as we roll this feature out - ie, w3infra can disable this requirement until we are ready to launch.