feat: add `subscription/list` capability (#1088)
Invoking `subscription/list` will retrieve a list of subscriptions for a
given account. The list of subscriptions allows us to discover spaces
(consumers) that are owned (paid for) by the account. With the list of
owned spaces we can make calls to `usage/report` to retrieve and display
usage information.