wip: introduce `update` method on `PlansStorage`
We already have a `set` method, but that is used to update our system when we get a webhook from Stripe. I debated adding an option to that or adding a new method to the interface specifically for a user of our system asking to update their plan.

Putting this PR up now to get feedback and thoughts - does anyone feel strongly that we should do this a different way?