feat: proof of concept of w3up-on-WASM
This is a proof-of-concept of running w3up on a WASM runtime.

Zinnia is "a sandboxed and resource-limited runtime for distributed workers" that
provides an easy way to run w3up on a WASM-based execution platform.

This proof-of-concept shows that it is possible to run w3up in "a sandboxed
and resource-limited runtime" which may be useful if we'd like to enable many different entities to run w3up instances.