fix: look for the correct capability in `client.capability.upload.get`
I was seeing some strange errors in the nft.storage work I'm doing right now and realized it's a result of a copypasta bug in `upload-client` where we query for the wrong capability when putting together this invocation.

We missed this in testing because tests use one big proof, which means that asking for `upload/add` is equivalent to asking for `upload/get` - in an ideal world I'd spend a bit of time untangling this but since I'm at the bottom of a stack of shaved yaks I'd like to get this shipped and file an issue to come back to this later.