Add blob index class Blobdex
Add a new class to represent a blob index. The blob index is a set of multihash, offset, and length values associated with the data stored in a blob.

The blob index in created when a client stores a blob of data. The multihash values in the blob index are indexed by IPNI if opting in to IPNI indexing. When a blob index is retrieved with a claims bundle, the client can use the items in the blob index to request specific data items, identified by multihash, from a stored blob.