fix!: storefront content store rename and separation for test (#1409)
With the aim to facilitate
https://github.com/w3s-project/w3up/issues/1349 storefront `dataStore`
is renamed to `contentStore` (note alignment of name with the `content`
property in Filecoin Pipeline capabilities/message namings). My goal is
to make `contentStore.stream()` perform HTTP Request to read thing from
Roundabout (if `filecoin/offer` comes with `content` CARCID code, it
will try to get a CAR, if `content` comes as RAW, it will try to get
RAW).

The main point in this change is to not require `contentStore` to have a
`put` method, like `dataStore` was requiring just for testing. It
extracts that requirement only to run tests, so that we can implement a
`contentStore` as something with just a `stream` function that behind
the scenes can perform a HTTP Request to Roundabout instead of being an
abstraction on top of a S3 bucket like today.

Note that this is not strictly needed, we could just use this as is and
in `dataStore` implementation and make `put`
https://github.com/w3s-project/w3infra/blob/main/filecoin/store/data.js#L39
`throw new Error('not implemented')` and just use an extended
`DataStore` class for w3infra testing. But I feel this way is nicer, and
also like more new naming as it aligns with everything including content
name defined in w3index
https://github.com/w3s-project/specs/blob/main/w3-index.md#sharded-dag-index-example

BREAKING CHANGE: dataStore in storefront renamed to contentStore