feat: add use cases package
Add an example of how to delegate all capabilities on a space to a `did:mailto` account. This has come up a few times recently and it's one potential solution to the "I want to change my email address" problem.

I'd like to start collecting use cases for w3up client here. It might be a good idea to run some or all of them as integration tests.