feat: Generate Space proofs on the fly, on `access/claim` (#1555)
The bulk of the work for
https://github.com/storacha/project-tracking/issues/125.

This together with [`w3ui` calling `access/claim` on page
load](https://github.com/storacha/w3ui/pull/644) will make new spaces
appear on the page without logging out.

---

## The Strategy

* On `access/confirm`, where we previously created a `*`/`ucan:*`
delegation with proofs for each space, we now create one with no proofs.
This delegation technically provides no capability, but serves as a
signal in `access/claim`.
* During `access/claim`, we look for the Agent's `*`/`ucan:*`
delegations in the store, and translate them on the fly into ones
containing proofs for the relevant spaces. This is legitimate, because
`*`/`ucan:*` means the Agent has been granted access to anything the
Account can do, it just doesn't have the proofs that Account can do
anything yet. We're just putting 2 and 2 together.

## Remaining

* [x] We don't validate the `*`/`ucan:*` delegation before recreating
it. If `access/confirm` doesn't create an attestation, `access/claim`
will still happily make one for its new delegation. That's a (minor?)
security issue. If the delegation is in our store, we can probably trust
it, but better to care about the attestation.
* [x] Tests for all of this.

---------

Co-authored-by: Alan Shaw <alan@storacha.network>