feat(w3up-client): add default gateway authorization (#1604)
### Context  
While updating the `web3-storage` docs, `w3cli`, and `@storacha/client`,
I found it inconvenient to repeatedly define the gateway DID and URL
every time I created a space. Also, if we don't provide the
`skipGatewayAuthorization=true`, it will throw an error saying it is
required to provide the gateway services to authorize.

### Changes
Following @travis suggestion (see:
https://github.com/storacha/docs/pull/21#discussion_r1883075639), I've
updated the client to automatically authorize the Storacha Gateway
(Production) by default if `skipGatewayAuthorization` is not set to
`true` and `authorizeGatewayServices` is not provided or is empty.

Once approved, this change will need to be ported to
[upload-service/w3up-client](https://github.com/storacha/upload-service/tree/main/packages/w3up-client).

And the PR https://github.com/storacha/freeway/pull/135 needs to be
merged before this change gets released.