fix: store w3s accept invocation (#142)
w3infra currently requires all receipts in agent messages to have
corresponding invocations. This PR simply adds the fake w3s accept
invocation for the (already added) w3s accept receipt to the agent
message so that this requirement can be met.

This also fixes an issue with registering blobs when a blob already
exists on a storage node.