feat: add `app` to `access/authorize` and `access/confirm` (#264)
We need a way to ensure users are redirected back to
`https://bsky.storage` after completing the Stripe checkout process.

To facilitate this, pass facts from `access/authorize`
to`access/confirm` - users can now add a fact like `{app:
'bsky-backups'}` to `access/authorize`, which will tell the
`access/confirm` handling code in `w3infra` to render the same pricing
table we use on https://bsky.storage which sends the user back to that
app after picking a plan via the email flow.

Add support for this to the client code via a new `appName` option in
the account creation functions that gets converted to a fact at the
lowest level.

There are a bunch of ways we could do this, but in the interest of
shipping ASAP I think this is a reasonable and fast way to go.

See https://github.com/storacha/w3infra/pull/472 for the other part of
this.