feat: port of https://github.com/storacha/upload-service/commit/2c12c23d13d14e9f1b79c34b8169f20975d431f7 (#1641)
Ports over the upload api changes in the above commit to upload service

-- original commit message
feat: add `app` to `access/authorize` and `access/confirm` (#264) (

We need a way to ensure users are redirected back to
`https://bsky.storage` after completing the Stripe checkout process.

To facilitate this, pass facts from `access/authorize`
to`access/confirm` - users can now add a fact like `{app:
'bsky-backups'}` to `access/authorize`, which will tell the
`access/confirm` handling code in `w3infra` to render the same pricing
table we use on https://bsky.storage which sends the user back to that
app after picking a plan via the email flow.

Add support for this to the client code via a new `appName` option in
the account creation functions that gets converted to a fact at the
lowest level.

There are a bunch of ways we could do this, but in the interest of
shipping ASAP I think this is a reasonable and fast way to go.

See https://github.com/storacha/w3infra/pull/472 for the other part of
this.